automatically set server in esp and don't turn on pump when ADC is out of range

+++ fw/bonsai_water.X/include/datastorage.h
@@ -21,7 +21,7 @@ struct STORAGE{
 typedef struct STORAGE storage_data;
 
 void storageAppendData( storage_data to_write );
-uint8_t storageGetData( uint8_t index, storage_data *data );
+uint8_t storageGetData( uint16_t index, storage_data *data );
 void storageClean();
 
 #endif
